Reorder tests to appropriate version
