+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/DefaultFileComparisonTestAssertionFailure.java
@@ -29,20 +29,11 @@ public class DefaultFileComparisonTestAssertionFailure extends DefaultTestAssert
     private final byte[] expectedContent;
     private final byte[] actualContent;
 
-    public DefaultFileComparisonTestAssertionFailure(@Nullable String message,
-                                                     String description,
-                                                     @Nullable String expected,
-                                                     @Nullable String actual,
-                                                     @Nullable List<? extends Failure> causes,
-                                                     String className,
-                                                     String stacktrace,
-                                                     @Nullable byte[] expectedContent,
-                                                     @Nullable byte[] actualContent) {
+    public DefaultFileComparisonTestAssertionFailure(String message, String description, String expected, String actual, List<? extends Failure> causes, String className, String stacktrace, byte[] expectedContent, byte[] actualContent) {
         super(message, description, expected, actual, causes, className, stacktrace);
         this.expectedContent = expectedContent;
         this.actualContent = actualContent;
     }
-
     @Nullable
     @Override
     public byte[] getExpectedContent() {
